fix(updater): match tauri-plugin-updater 2.10 progress callback signature
tauri-plugin-updater 2.10.1 changed the download progress callback's
first argument from u64 to usize. The progress closure in updater.rs
still declared u64, so the Tauri release build failed to compile with a
closure type mismatch. `downloaded` is only used via `as f64`,
`saturating_sub`, and logging, so usize is a drop-in.

+++ src-tauri/src/updater.rs
@@ -82,7 +82,7 @@ async fn run_one_check(app: &AppHandle) {
             // 100MB install. Metered-network detection is a known gap (see
             // privacy contract above); platform-specific NWPathMonitor /
             // Windows NetworkInformation integration is tracked separately.
-            let progress = |downloaded: u64, total: Option<u64>| {
+            let progress = |downloaded: usize, total: Option<u64>| {
                 if let Some(total) = total.filter(|&t| t > 0) {
                     let curr_pct = (downloaded as f64 / total as f64 * 100.0) as u32;
                     let prev = downloaded.saturating_sub(1);
